--- squad-2-selbetti-1/Codenation.ErrorCenter/Codenation.ErrorCenter.Models/Models/ErrorCenterContext.cs
@@ -1,16 +1,0 @@
-﻿using Codenation.ErrorCenter.Models.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace Codenation.ErrorCenter.Models.Models
-{
-    public class ErrorCenterContext : DbContext
-    {
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            optionsBuilder.UseSqlServer(@"Server=.;Database=errorLog;Trusted_Connection=True");
-        }
-
-        public DbSet<Log> Logs { get; set; }
-        public DbSet<User> Users { get; set; }
-    }
-}
